Move search code under service
